Advanced Haskell solution
Added an advanced solution in Haskell. Yes, I've decided to learn Haskell. It makes me feel like a wizard! It took me 50+ lines and several hours to partially solve this problem two years ago, now I can solve it in an hour in a language I've known for a week!